botlib: C++ compatibility for l_precomp

+++ code/botlib/l_precomp.c
@@ -2485,7 +2485,7 @@ static int PC_Directive_evalfloat(source_t *source)
 	token.whitespace_p = source->scriptstack->script_p;
 	token.endwhitespace_p = source->scriptstack->script_p;
 	token.linescrossed = 0;
-	sprintf(token.string, "%1.2f", fabs(value));
+	sprintf(token.string, "%1.2f", fabs((double)(value)));
 	token.type = TT_NUMBER;
 	token.subtype = TT_FLOAT|TT_LONG|TT_DECIMAL;
 	PC_UnreadSourceToken(source, &token);
@@ -2598,12 +2598,12 @@ static int PC_DollarDirective_evalfloat(source_t *source)
 	token.whitespace_p = source->scriptstack->script_p;
 	token.endwhitespace_p = source->scriptstack->script_p;
 	token.linescrossed = 0;
-	sprintf(token.string, "%1.2f", fabs(value));
+	sprintf(token.string, "%1.2f", fabs((double)(value)));
 	token.type = TT_NUMBER;
 	token.subtype = TT_FLOAT|TT_LONG|TT_DECIMAL;
 
 #ifdef NUMBERVALUE
-	token.floatvalue = fabs(value);
+	token.floatvalue = fabs((double)(value));
 	token.intvalue = (unsigned long) token.floatvalue;
 #endif //NUMBERVALUE
